+++ bootique-swagger-ui/src/test/java/io/bootique/swagger/ui/SwaggerUiDefaultLocationIT.java
@@ -46,7 +46,7 @@ public class SwaggerUiDefaultLocationIT extends SwaggerUiBaseIT {
 
     @Test
     public void testApi_Console() {
-        Response r = BASE_TARGET.path("/swagger").request().get();
+        Response r = BASE_TARGET.path("/swagger-ui").request().get();
         assertEquals(200, r.getStatus());
         assertEqualsToResourceContents("swagger-response1.html", r.readEntity(String.class));
     }